training/requirements-caption.txt
transformers>=4.57.0
accelerate>=1.10.0
qwen-vl-utils>=0.0.14
safetensors>=0.4.5
pillow>=11.0.0
